fix(clinical) move schema to file
